app/services/image_processing_service.py
from typing import Dict, Any
from io import BytesIO
from PIL import Image
import base64

class ImageProcessingService:
    @staticmethod
    def process_image_file(file_content: bytes) -> Dict[str, Any]:
        try:
            image = Image.open(BytesIO(file_content))
            image.verify() # Verify that this is indeed an image

            # For now, we'll just return the image as a base64 encoded string
            # In a real application, you would perform OCR, object detection, etc.
            buffered = BytesIO()
            image.save(buffered, format="PNG")
            base64_image = base64.b64encode(buffered.getvalue()).decode('utf-8')
            
            return {
                "image_data_base64": base64_image,
                "width": image.width,
                "height": image.height,
                "format": image.format,
                "mode": image.mode
            }
        except Exception as e:
            print(f"Error processing image: {e}")
            return {"error": str(e)}
