requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.0
email-validator>=2.0.0
pyyaml==6.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pandas==2.1.3
openpyxl>=3.0.0
pypdf==3.17.4
Pillow==10.1.0
rapidfuzz==3.6.1
